added tags support
