+++ package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@cucumber/cucumber-expressions": "^18.0.1",
     "@cucumber/gherkin": "^35.1.0",
-    "@cucumber/tag-expressions": "^6.2.0",
+    "@cucumber/tag-expressions": "^7.0.0",
     "@cypress/webpack-preprocessor": "^7.0.1",
     "fs-extra": "^11.3.2"
   },
