+++ test/cypress.config.js
@@ -6,7 +6,7 @@ module.exports = defineConfig({
     specPattern: 'test/cypress/feature/**/*.feature',
     supportFile: 'test/cypress/support/e2e.js',
     setupNodeEvents(on, config) {
-      on('file:preprocessor', cucumber)
+      on('file:preprocessor', cucumber);
     },
     viewportWidth: 1920,
     viewportHeight: 1080,
