+++ test/cypress/support/parameter_types.js
@@ -10,4 +10,4 @@ defineParameterType({
     name: 'color',
     regexp: /(red|blue|green)/,
     transformer: color => ({ color, hex: hexes[color] })
-})
+});
